+++ src/main/kotlin/minesweeper/domain/gameboard/MineBoard.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.gameboard
 
 import kotlin.random.Random
 
@@ -7,7 +7,11 @@ class MineBoard(
     val height: Int,
     mineNumber: Int
 ) {
-    val board: Array<Row> = Array(height) { Row(width) }
+    val board: Array<Row> = Array(height) {
+        Row(
+            width
+        )
+    }
 
     init {
         require(mineNumber < width * height) { "마인의 개수가 너무 많습니다." }
